pyproject.toml
[project]
name = "user-profile-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=6.3.0",
    "motor>=3.7.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "tqdm>=4.67.1",
    "google-genai>=1.0.0",
    "httpx[socks]>=0.27.0",
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
]
